CLUE-576 Move MetadataDocMapModel next to DocumentMetadataModel
DocumentMetadataStore imported MetadataDocMapModel from sorted-documents while
sorted-documents imported the store, forming an import cycle. The map model is
just types.map(DocumentMetadataModel), so its natural home is beside that model;
moving it there makes the store -> sorted-documents edge disappear and leaves a
single direction of dependency.
