+++ src/models/stores/document-group.test.ts
@@ -9,7 +9,8 @@ import { DocumentContentSnapshotType } from "../document/document-content";
 import { GroupDocument, ProblemDocument } from "../document/document-types";
 import { ClassModel, ClassModelType, ClassUserModel } from "./class";
 import { GroupModel, GroupsModel, GroupsModelType, GroupUserModel } from "./groups";
-import { ISortedDocumentsStores, MetadataDocMapModel, SortedDocuments } from "./sorted-documents";
+import { MetadataDocMapModel } from "../document/document-metadata-model";
+import { ISortedDocumentsStores, SortedDocuments } from "./sorted-documents";
 import { DocumentMetadataStore } from "./document-metadata-store";
 import { DB } from "../../lib/db";
 import { Bookmark, Bookmarks } from "./bookmarks";
